+++ src/app/router/index.ts
@@ -1,0 +1,3 @@
+export * from './routes';
+export * from './routes.actions';
+export * from './routes.effects';
